+++ mountpoint-s3-client/src/s3_crt_client.rs
@@ -37,8 +37,8 @@ use tracing::{debug, error, trace, Span};
 
 use self::get_object::S3GetObjectRequest;
 use self::put_object::S3PutObjectRequest;
-use crate::endpoint_config::EndpointConfig;
 use crate::endpoint_config::EndpointError;
+use crate::endpoint_config::{self, EndpointConfig};
 use crate::object_client::*;
 use crate::user_agent::UserAgent;
 
@@ -303,6 +303,22 @@ impl S3CrtClientInner {
             None,
             None,
         );
+
+        let endpoint_config = match endpoint_config.get_endpoint() {
+            None => {
+                // No explicit endpoint was configured, let's try the environment variable
+                if let Ok(aws_endpoint_url) = std::env::var("AWS_ENDPOINT_URL") {
+                    debug!("using AWS_ENDPOINT_URL {}", aws_endpoint_url);
+                    let env_uri = Uri::new_from_str(&allocator, &aws_endpoint_url)
+                        .map_err(|e| EndpointError::InvalidUri(endpoint_config::InvalidUriError::CouldNotParse(e)))?;
+                    endpoint_config.endpoint(env_uri)
+                } else {
+                    endpoint_config
+                }
+            }
+            Some(_) => endpoint_config,
+        };
+
         client_config.express_support(true);
         client_config.signing_config(signing_config);
 
@@ -1087,6 +1103,7 @@ impl ObjectClient for S3CrtClient {
 #[cfg(test)]
 mod tests {
     use mountpoint_s3_crt::common::error::Error;
+    use rusty_fork::rusty_fork_test;
     use std::assert_eq;
 
     use super::*;
@@ -1153,6 +1170,63 @@ mod tests {
         assert!(user_agent_header_value
             .to_string_lossy()
             .starts_with(expected_user_agent));
+    }
+
+    fn assert_expected_host(expected_host: &str, endpoint_config: EndpointConfig) {
+        let config = S3ClientConfig {
+            endpoint_config,
+            ..Default::default()
+        };
+
+        let client = S3CrtClient::new(config).expect("create test client");
+
+        let mut message = client
+            .inner
+            .new_request_template("GET", "")
+            .expect("new request template expected");
+
+        let headers = message.inner.get_headers().expect("expected a block of HTTP headers");
+
+        let host_header = headers.get("Host").expect("Host header expected");
+        let host_header_value = host_header.value();
+
+        assert_eq!(host_header_value.to_string_lossy(), expected_host);
+    }
+
+    // run with rusty_fork to avoid issues with other tests and their env variables.
+    rusty_fork_test! {
+        #[test]
+        fn test_endpoint_favors_parameter_over_env_variable() {
+            let endpoint_uri = Uri::new_from_str(&Allocator::default(), "https://s3.us-west-2.amazonaws.com").unwrap();
+            let endpoint_config = EndpointConfig::new("region-place-holder").endpoint(endpoint_uri);
+            std::env::set_var("AWS_ENDPOINT_URL", "https://s3.us-east-1.amazonaws.com");
+            // even though we set the environment variable, the parameter takes precedence
+            assert_expected_host("s3.us-west-2.amazonaws.com", endpoint_config);
+        }
+
+        #[test]
+        fn test_endpoint_favors_env_variable() {
+            let endpoint_config = EndpointConfig::new("us-east-1");
+            std::env::set_var("AWS_ENDPOINT_URL", "https://s3.eu-west-1.amazonaws.com");
+            assert_expected_host("s3.eu-west-1.amazonaws.com", endpoint_config);
+        }
+
+        #[test]
+        fn test_endpoint_with_invalid_env_variable() {
+            let endpoint_config = EndpointConfig::new("us-east-1");
+            std::env::set_var("AWS_ENDPOINT_URL", "htp:/bad:url");
+            let config = S3ClientConfig {
+                endpoint_config,
+                ..Default::default()
+            };
+
+            let client = S3CrtClient::new(config);
+            match client {
+                Ok(_) => panic!("expected an error"),
+                Err(e) => assert_eq!(e.to_string().to_lowercase(), "invalid s3 endpoint"),
+            }
+        }
+
     }
 
     /// Simple test to ensure the user agent header is correct even when prefix is not added