+++ mountpoint-s3/tests/common/mod.rs
@@ -107,3 +107,9 @@ fn init_tracing_subscriber() {
     let _ = RustLogAdapter::try_init();
     let _ = tracing_subscriber::fmt::try_init();
 }
+
+#[ctor::ctor]
+fn init_crt() {
+    mountpoint_s3_crt::io::io_library_init(&mountpoint_s3_crt::common::allocator::Allocator::default());
+    mountpoint_s3_crt::s3::s3_library_init(&mountpoint_s3_crt::common::allocator::Allocator::default());
+}